+++ webbangiaydep/src/main/java/business/AddToCart/AddToCartInputData.java
@@ -3,14 +3,14 @@ package business.AddToCart;
 
 public class AddToCartInputData {
     
-    public AddToCartInputData(Long userId, int productId, int variantId, int quantity) {
+    public AddToCartInputData(Long userId, Long productId, Long variantId, int quantity) {
         this.userId = userId;
         this.productId = productId;
         this.variantId = variantId;
         this.quantity = quantity;
     }
     public Long userId;
-    public int productId;
-    public int variantId;
+    public Long productId;
+    public Long variantId;
     public int quantity;
 }
