+++ webbangiaydep/src/main/java/persistence/ViewShoeCart/ViewShoeCartDTO.java
@@ -3,8 +3,8 @@ package persistence.ViewShoeCart;
 public class ViewShoeCartDTO 
 {
     public int cartId;       
-    public int productId;
-    public int variantId;    
+    public Long productId;
+    public Long variantId;    
     public String productName;
     public String imageUrl;
     public int size;
