+++ webbangiaydep/src/main/java/persistence/ViewShoeList/ViewShoeListDTO.java
@@ -2,7 +2,7 @@ package persistence.ViewShoeList;
 
 public class ViewShoeListDTO 
 {
-    public int id;           // Mã giày
+    public Long id;           // Mã giày
     public String name;      // Tên giày
     public double price;     // Giá bán
     public String description;
